make common tests.accessories
